Finishing logic and debugging with mealplans. Refactored and finished setting up ShoppingList. Realized I should have switched branches midway... Issues #16, 18, 19

+++ app/src/main/java/com/example/happymeals/ingredient/IngredientStorage.java
@@ -78,7 +78,7 @@ public class IngredientStorage implements DatabaseListener {
      * @param count The {@link Double} representing the count to take away from the
      *              provided {@link Ingredient}.
      */
-    public void requestConsumptionOfIngredient( Ingredient ingredient, int count ) {
+    public void requestConsumptionOfIngredient( Ingredient ingredient, Double count ) {
         ingredient.setAmount( ingredient.getAmount() - count );
         updateIngredient( ingredient );
     }
